[raffle] script
