+++ raffle/inputs/parties3.json
@@ -1,0 +1,13 @@
+[
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm",
+    "addr_test1vrfgr2dny4zsa45xc5e0wl6xpv64tlzjzkw8znzsnmu87vc86pndm"
+]
