+++ raffle/inputs/prizes.json
@@ -1,0 +1,6 @@
+[
+    [
+        "89cd4d46ee5982a760cd12886224fbb285aa98211f19d354ccbe452b",
+        "1stPrize"
+    ]
+]
