+++ raffle/outputs/unsigned-tx.txt
@@ -1,0 +1,5 @@
+{
+    "type": "TxBodyBabbage",
+    "description": "",
+    "cborHex": "86a40081825820a91bcbeb910ba2ac754a69d626a21c7a19cf51b921306c1014916b2435cd834e000182a200583900e841123c552664b31bc2f24447967508737b9bb4dc88fb577d6f3f598b461c30918a0417b49964ad62c6fda538fbbac94895c106c0147554011b000000058a1a9a68a300581d702ed2631dbb277c84334453c5c437b86325d371f0835a28b910a91a6e011a001e8480028200582064ffcac0ffed9d258d24611414043aa5e62d7ead54eb565d1889881a777e5a5d021a0002ea010b5820c09da70d883a3dced42801e116e135190e6c37fca6ddc0e5f0aa78d9073e08ac9fff81d8799fd8799f40ffd8799fa1d8799fd8799fd87980d8799fd8799f581ce841123c552664b31bc2f24447967508737b9bb4dc88fb577d6f3f59ffd8799fd8799fd8799f581c8b461c30918a0417b49964ad62c6fda538fbbac94895c106c0147554ffffffffffd8799f4040ffff1a001e8480a0a000ffd87c9f9fd87a9fd8799fd8799fd87980d8799fd8799f581ce841123c552664b31bc2f24447967508737b9bb4dc88fb577d6f3f59ffd8799fd8799fd8799f581c8b461c30918a0417b49964ad62c6fda538fbbac94895c106c0147554ffffffffffd8799fd87980d8799fd8799f581ce841123c552664b31bc2f24447967508737b9bb4dc88fb577d6f3f59ffd8799fd8799fd8799f581c8b461c30918a0417b49964ad62c6fda538fbbac94895c106c0147554ffffffffffd8799f581c89cd4d46ee5982a760cd12886224fbb285aa98211f19d354ccbe452b483173745072697a65ffd87a9f01ffff58206b3ada29df1182e8747967f9d9b22569163b5acbeda899214fc5a7c5f8c15843ffff1b00000189bb6f8986d87980ffff80f5f6"
+}
